requirements.txt
# Core config and data models
pyyaml>=6.0
pydantic>=2.0
python-dotenv>=1.0

# LLM providers
anthropic>=0.34.0
openai>=1.40.0

# Retrieval and embeddings
sentence-transformers>=3.0.0
rank-bm25>=0.2.2

# Data processing and scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
markdownify>=0.13.1

# Numeric/evaluation utilities
numpy>=1.26.0
matplotlib>=3.8.0
tqdm>=4.66.0

# Optional (vector DB backend)
# Install when you want Chroma vector search instead of BM25-only fallback.
chromadb==0.4.24 ; platform_system != "Windows"

